Revert header changes
