+++ src/openrct2/peep/Peep.h
@@ -17,8 +17,8 @@
 #include "../ride/RideTypes.h"
 #include "../ride/ShopItem.h"
 #include "../util/Util.h"
-#include "../world/Entity.h"
 #include "../world/Location.hpp"
+#include "../world/SpriteBase.h"
 
 #include <algorithm>
 #include <bitset>
