+++ src/openrct2/ride/Vehicle.h
@@ -13,8 +13,8 @@
 #include "../audio/audio.h"
 #include "../common.h"
 #include "../ride/RideTypes.h"
-#include "../world/Entity.h"
 #include "../world/Location.hpp"
+#include "../world/SpriteBase.h"
 #include "Station.h"
 #include "VehicleColour.h"
 #include "VehicleEntry.h"
