+++ src/openrct2/world/Entity.h
@@ -13,24 +13,6 @@
 
 constexpr uint16_t MAX_ENTITIES = 10000;
 
-enum class EntityType : uint8_t
-{
-    Vehicle,
-    Guest,
-    Staff,
-    Litter,
-    SteamParticle,
-    MoneyEffect,
-    CrashedVehicleParticle,
-    ExplosionCloud,
-    CrashSplash,
-    ExplosionFlare,
-    JumpingFountain,
-    Balloon,
-    Duck,
-    Count
-};
-
 SpriteBase* try_get_sprite(size_t spriteIndex);
 SpriteBase* get_sprite(size_t sprite_idx);
