+++ src/openrct2/world/SpriteBase.h
@@ -4,7 +4,24 @@
 
 struct CoordsXYZ;
 enum class SpriteIdentifier : uint8_t;
-enum class EntityType : uint8_t;
+
+enum class EntityType : uint8_t
+{
+    Vehicle,
+    Guest,
+    Staff,
+    Litter,
+    SteamParticle,
+    MoneyEffect,
+    CrashedVehicleParticle,
+    ExplosionCloud,
+    CrashSplash,
+    ExplosionFlare,
+    JumpingFountain,
+    Balloon,
+    Duck,
+    Count
+};
 
 struct SpriteBase
 {
